# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/product-images": {
            "get": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Get list of all product images",
                "operationId": "7d239e28af79ac485e7147b5cfc502ae",
                "responses": {
                    "200": {
                        "description": "List of product images",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductImage"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Create a new product image",
                "operationId": "6651e900a3d851d935af58840f61e47d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "product_id",
                                    "path"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "image1.jpg"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "path": {
                                        "type": "string",
                                        "example": "uploads/images/image1.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product image created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductImage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/product-images/{id}": {
            "get": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Get a specific product image by ID",
                "operationId": "349756bcbde0c7049f725277c79f4359",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product image ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product image data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductImage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product image not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Update a product image",
                "operationId": "8b3ebce8589d1b390bb0137ad7b48ee5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product image ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "updated_image.jpg"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "path": {
                                        "type": "string",
                                        "example": "uploads/updated_image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product image updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductImage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product image not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Delete a product image",
                "operationId": "25c20db020594e10c511737a06f719e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product image ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product image deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "ProductImage deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product image not found"
                    }
                }
            }
        },
        "/products/{product_id}/images": {
            "get": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Get all images for a specific product",
                "operationId": "4e638494c76c102876a1608922684eb4",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product images",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductImage"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductImage": {
                "title": "Product Image",
                "description": "Gambar tambahan untuk produk",
                "required": [
                    "product_images_id",
                    "name",
                    "product_id",
                    "path"
                ],
                "properties": {
                    "product_images_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "product_image_1.jpg"
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "path": {
                        "type": "string",
                        "example": "path/to/image.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-23T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-23T12:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}